navbar file moved to components
